# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/profiles/me": {
      "get": {
        "operationId": "ProfilesController_getMyProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get current user profile with completeness score",
        "tags": ["profiles"]
      },
      "put": {
        "operationId": "ProfilesController_updateMyProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateProfileDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Profile" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update current user profile",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/cv": {
      "post": {
        "operationId": "ProfilesController_uploadCv",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": { "file": { "type": "string", "format": "binary" } }
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Upload CV and auto-populate profile with AI parsing",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/avatar": {
      "post": {
        "operationId": "ProfilesController_uploadAvatar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": { "file": { "type": "string", "format": "binary" } }
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Upload profile avatar",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/visibility": {
      "put": {
        "operationId": "ProfilesController_updateVisibility",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateVisibilityDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Profile" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update profile visibility settings",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/completeness": {
      "get": {
        "operationId": "ProfilesController_getCompleteness",
        "parameters": [
          {
            "name": "targetRole",
            "required": false,
            "in": "query",
            "description": "Target role to calculate completeness for",
            "schema": { "example": "Backend Developer", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get profile completeness score for target role",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/cv/sessions": {
      "get": {
        "operationId": "ProfilesController_getCvSessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CvImportSession" }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get CV import sessions for current user",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/cv/confirm/{sessionId}": {
      "post": {
        "operationId": "ProfilesController_confirmCvImport",
        "parameters": [
          { "name": "sessionId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Confirm CV import session and merge to profile",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/cv/discard/{sessionId}": {
      "post": {
        "operationId": "ProfilesController_discardCvImport",
        "parameters": [
          { "name": "sessionId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Discard CV import session",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/insights": {
      "get": {
        "operationId": "ProfilesController_getInsights",
        "parameters": [
          {
            "name": "unreadOnly",
            "required": false,
            "in": "query",
            "schema": { "type": "boolean" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ProfileInsight" }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get AI insights for current user profile",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/insights/{insightId}/read": {
      "post": {
        "operationId": "ProfilesController_markInsightAsRead",
        "parameters": [
          { "name": "insightId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProfileInsight" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Mark insight as read",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/insights/{insightId}/actioned": {
      "post": {
        "operationId": "ProfilesController_markInsightAsActioned",
        "parameters": [
          { "name": "insightId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProfileInsight" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Mark insight as actioned",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/skills": {
      "get": {
        "operationId": "ProfilesController_getSkills",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ProfileSkill" }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get all skills for current user",
        "tags": ["profiles"]
      },
      "post": {
        "operationId": "ProfilesController_createSkill",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateSkillDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProfileSkill" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Add a new skill",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/skills/{skillId}": {
      "put": {
        "operationId": "ProfilesController_updateSkill",
        "parameters": [
          { "name": "skillId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateSkillDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProfileSkill" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update a skill",
        "tags": ["profiles"]
      },
      "delete": {
        "operationId": "ProfilesController_deleteSkill",
        "parameters": [
          { "name": "skillId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Delete a skill",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/skills/merge": {
      "post": {
        "operationId": "ProfilesController_mergeSkills",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/MergeSkillsDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProfileSkill" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Merge duplicate skills",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/experiences": {
      "get": {
        "operationId": "ProfilesController_getExperiences",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ProfileExperience" }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get all experiences for current user",
        "tags": ["profiles"]
      },
      "post": {
        "operationId": "ProfilesController_createExperience",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateExperienceDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProfileExperience" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Add a new experience",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/experiences/{experienceId}": {
      "put": {
        "operationId": "ProfilesController_updateExperience",
        "parameters": [
          { "name": "experienceId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateExperienceDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProfileExperience" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update an experience",
        "tags": ["profiles"]
      },
      "delete": {
        "operationId": "ProfilesController_deleteExperience",
        "parameters": [
          { "name": "experienceId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Delete an experience",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/projects": {
      "get": {
        "operationId": "ProfilesController_getProjects",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ProfileProject" }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get all projects for current user",
        "tags": ["profiles"]
      },
      "post": {
        "operationId": "ProfilesController_createProject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateProjectDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProfileProject" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Add a new project",
        "tags": ["profiles"]
      }
    },
    "/api/profiles/me/projects/{projectId}": {
      "put": {
        "operationId": "ProfilesController_updateProject",
        "parameters": [
          { "name": "projectId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateProjectDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProfileProject" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update a project",
        "tags": ["profiles"]
      },
      "delete": {
        "operationId": "ProfilesController_deleteProject",
        "parameters": [
          { "name": "projectId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Delete a project",
        "tags": ["profiles"]
      }
    },
...
  },
  "info": {
    "title": "AI Job Application Context API",
    "description": "The AI Job Application API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "fullName": { "type": "string", "description": "Full name", "example": "Nguyen Van A" },
          "phone": { "type": "string", "description": "Phone number", "example": "+84912345678" },
          "address": { "type": "string", "description": "Address", "example": "Ha Noi, Vietnam" },
          "currentRole": {
            "type": "string",
            "description": "Current job role",
            "example": "Senior Frontend Developer"
          },
          "seniorityLevel": {
            "enum": ["entry", "junior", "mid", "senior", "lead", "manager"],
            "type": "string",
            "description": "Seniority level"
          },
          "yearsOfExperience": {
            "type": "number",
            "minimum": 0,
            "maximum": 50,
            "description": "Years of experience",
            "example": 5
          },
          "location": {
            "type": "string",
            "description": "Current location",
            "example": "Ho Chi Minh City"
          },
          "workPreference": {
            "enum": ["remote", "onsite", "hybrid", "flexible"],
            "type": "string",
            "description": "Preferred work mode"
          },
          "linkedin": {
            "type": "string",
            "description": "LinkedIn profile URL",
            "example": "https://linkedin.com/in/johndoe"
          },
          "portfolio": {
            "type": "string",
            "description": "Portfolio website URL",
            "example": "https://johndoe.dev"
          }
        }
      },
...
    }
  }
}